requirements.txt
appengine-python-standard>=1.0.0
Flask==2.1.0
google-cloud-storage
google-cloud-datastore==2.7.1

